Add hook for allowance state
