+++ src/pages/SplitPosition/Form.tsx
@@ -1,5 +1,6 @@
 import { useWeb3Connected } from 'contexts/Web3Context'
 import { BigNumber } from 'ethers/utils'
+import { useAllowanceState } from 'hooks/useAllowanceState'
 import React, { useCallback, useEffect, useMemo, useState } from 'react'
 import { useForm } from 'react-hook-form'
 import { ERC20Service } from 'services/erc20'
@@ -18,7 +19,6 @@ import { Row } from '../../components/pureStyledComponents/Row'
 import { TitleControl } from '../../components/pureStyledComponents/TitleControl'
 import { TitleValue } from '../../components/text/TitleValue'
 import { NULL_PARENT_ID, ZERO_BN } from '../../config/constants'
-import { Remote } from '../../util/remoteData'
 import { trivialPartition } from '../../util/tools'
 import { Token } from '../../util/types'
 
@@ -43,8 +43,6 @@ export type SplitPositionFormMethods = {
 }
 
 interface Props {
-  allowance: Remote<BigNumber>
-  unlockCollateral: () => void
   onCollateralChange: (collateral: string) => void
   splitPosition: (
     collateral: string,
@@ -53,18 +51,10 @@ interface Props {
     partition: BigNumber[],
     amount: BigNumber
   ) => void
-  hasUnlockedCollateral: boolean
   tokens: Token[]
 }
 
-export const Form = ({
-  allowance,
-  hasUnlockedCollateral,
-  onCollateralChange,
-  splitPosition,
-  tokens,
-  unlockCollateral,
-}: Props) => {
+export const Form = ({ onCollateralChange, splitPosition, tokens }: Props) => {
   const DEFAULT_VALUES = useMemo(() => {
     return {
       conditionId: '',
@@ -160,21 +150,12 @@ export const Form = ({
     signer,
   ])
 
-  const hasEnoughAllowance = allowance.map(
-    (allowance) => allowance.gte(amount) && !allowance.isZero()
+  const { allowanceFinished, fetching, showAskAllowance, unlockCollateral } = useAllowanceState(
+    collateralToken.address,
+    amount
   )
 
-  const notEnoughAllowance = hasEnoughAllowance.hasData() && !hasEnoughAllowance.get()
-
-  console.log('successWithEnoughAllowance', notEnoughAllowance)
-
-  // We show the allowance component if
-  // - *We know* that the user doesn't have enough allowance
-  // - The user just unlocked his collateral
-  // - allowance is loading
-  const showAskAllowance = notEnoughAllowance || hasUnlockedCollateral || allowance.isLoading()
-
-  const canSubmit = isValid && (hasEnoughAllowance.getOr(false) || hasUnlockedCollateral)
+  const canSubmit = isValid && allowanceFinished
   const mockedNumberedOutcomes = [
     [1, 4, 3],
     [6, 5],
@@ -206,8 +187,8 @@ export const Form = ({
       {showAskAllowance && (
         <SetAllowance
           collateral={collateralToken}
-          fetching={allowance.isLoading()}
-          finished={hasUnlockedCollateral && hasEnoughAllowance.getOr(false)}
+          fetching={fetching}
+          finished={allowanceFinished}
           onUnlock={unlockCollateral}
         />
       )}
